Add timeout to go back N algorithm and receive only the byte_array instead of Vector<Header>

+++ src/config.rs
@@ -34,7 +34,7 @@ pub const NODES: Option<&[Node]> = None;
 
 // Configurações da comunicação
 pub const W_SIZE: usize = 5;
-pub const TIMEOUT: u128 = 100; // 0.1 segundo
+pub const TIMEOUT: u64 = 100; // 0.1 segundo
 pub const HEARTBEAT_INTERVAL: u64 = 500; // 500 milissegundos
 pub const FAILURE_DETECTION_INTERVAL: u64 = 1000; // 1 segundo
 pub const BUFFER_SIZE: usize = 1024;
